Begin support for "hefer" host configuration
This commit makes some initial changes to support the configuration of a
new host named "hefer". This includes the creation of two new roles:

  - foundation
  - xubuntu

The configuration found in these roles was pulled out of the "reptar"
specific role, such that they could easily be used to configure both
"hefer" and "reptar".